moves models.py from internal upper into dto hierarchy
